Added UF2 formatted firmware
This firmware can be loaded onto Pico W by holding the BOOTSEL button while powering on and then copying the file to the removal storage that it presents. Upon booting, it will wait several seconds for a USB serial connection at baud 115200. It will then boot FreeRTOS and attempt to connect to a WIFI SSID `DemoNet` with password `Password123!`